move from draw-map.js to zoomPanCalcs.js : ZoomFilter : wheelDelta(), zoomFilter()
move from draw-map.js to zoomPanCalcs.js : added ZoomFilter() : wheelDelta(), zoomFilter().
